feat(signers-page): added unit tests
