+++ src/app/signers/SignersHeader.tsx
@@ -18,7 +18,8 @@ import { ExplorerErrorBoundary } from '../_components/ErrorBoundary';
 import { useSuspenseCurrentStackingCycle } from '../_components/Stats/CurrentStackingCycle/useCurrentStackingCycle';
 import { useSuspenseNextStackingCycle } from '../_components/Stats/NextStackingCycle/useNextStackingCycle';
 import { CurrentCycleCard } from './CurrentCycle';
-import { SignerDistributionHeader, SignersDistribution } from './SignerDistribution';
+import { SignersDistribution } from './SignerDistribution';
+import { SignerDistributionHeader } from './SignerDistributionHeader';
 import { SignersStackersData, useGetStackersBySignerQuery } from './data/UseSignerAddresses';
 import { useStxSupply } from './data/usStxSupply';
 import { useSuspensePoxSigners } from './data/useSigners';
